requirements.txt
openai
transformers
torch
python-dotenv
openai
matplotlib